feat(agentos): extract AgentIntentionGenerator.kt from AgentAdvanced (WZ-31597)
