+++ agentos/agentos-service/src/main/kotlin/io/whozoss/agentos/agent/AgentIntentionGenerationException.kt
@@ -1,0 +1,3 @@
+package io.whozoss.agentos.agent
+
+internal class AgentIntentionGenerationException(message: String) : Exception(message)
