+++ agentos/agentos-service/src/test/kotlin/io/whozoss/agentos/agent/AgentAdvancedSpec.kt
@@ -1,6 +1,5 @@
 package io.whozoss.agentos.agent
 
-import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.shouldBe
@@ -23,132 +22,34 @@ class AgentAdvancedSpec :
     StringSpec({
         timeout = 5000
 
-        // -------------------------------------------------------------------------
-        // Parsing unit tests — no ChatClient involved
-        // -------------------------------------------------------------------------
-
         fun makeParserAgent(): AgentAdvanced {
             val mockChatClient = mockk<ChatClient>(relaxed = true)
+            val agentId = UUID.randomUUID()
+            val context =
+                AgentAdvancedContext(
+                    chatClient = mockChatClient,
+                    tools = emptyList(),
+                    instructions = null,
+                    agentId = agentId,
+                )
             return AgentAdvanced(
                 name = "ParserAgent",
-                chatClient = mockChatClient,
-                tools = emptyList(),
+                context = context,
+                intentionGenerator = mockk(),
             )
         }
 
-        val validTools = listOf("FILES__ReadFile", "JIRA__GetIssue", "Answer")
-
-        "parseIntentionAndTool — nominal XML format" {
-            val agent = makeParserAgent()
-            val response =
-                """
-                <intention>I need to read the file to answer the question.</intention>
-                <toolName>FILES__ReadFile</toolName>
-                """.trimIndent()
-
-            val (intention, toolName) = agent.parseIntentionAndTool(response, validTools)
-
-            intention shouldBe "I need to read the file to answer the question."
-            toolName shouldBe "FILES__ReadFile"
-        }
-
-        "parseIntentionAndTool — tags on a single line" {
-            val agent = makeParserAgent()
-            val response = "<intention>Done.</intention><toolName>Answer</toolName>"
-
-            val (intention, toolName) = agent.parseIntentionAndTool(response, validTools)
-
-            intention shouldBe "Done."
-            toolName shouldBe "Answer"
-        }
-
-        "parseIntentionAndTool — extra text outside tags is ignored" {
-            val agent = makeParserAgent()
-            val response =
-                """
-                Here is my response:
-                <intention>Fetching the Jira issue for context.</intention>
-                <toolName>JIRA__GetIssue</toolName>
-                Let me know if you need more.
-                """.trimIndent()
-
-            val (intention, toolName) = agent.parseIntentionAndTool(response, validTools)
-
-            intention shouldBe "Fetching the Jira issue for context."
-            toolName shouldBe "JIRA__GetIssue"
-        }
-
-        "parseIntentionAndTool — multi-line intention content" {
-            val agent = makeParserAgent()
-            val response =
-                """
-                <intention>
-                Step 1: check state.
-                Step 2: call the tool.
-                </intention>
-                <toolName>FILES__ReadFile</toolName>
-                """.trimIndent()
-
-            val (intention, toolName) = agent.parseIntentionAndTool(response, validTools)
-
-            intention shouldContain "Step 1"
-            intention shouldContain "Step 2"
-            toolName shouldBe "FILES__ReadFile"
-        }
-
-        "parseIntentionAndTool — unknown tool name falls back to Answer" {
-            val agent = makeParserAgent()
-            val response =
-                """
-                <intention>Trying a non-existent tool.</intention>
-                <toolName>UNKNOWN__Tool</toolName>
-                """.trimIndent()
-
-            val (_, toolName) = agent.parseIntentionAndTool(response, validTools)
-
-            toolName shouldBe "Answer"
-        }
-
-        "parseIntentionAndTool — missing toolName tag throws IntentionParsingException" {
-            val agent = makeParserAgent()
-            val response = "<intention>No tool tag present.</intention>"
-
-            shouldThrow<AgentAdvanced.IntentionParsingException> {
-                agent.parseIntentionAndTool(response, validTools)
-            }
-        }
-
-        "parseIntentionAndTool — missing intention tag uses full response as intention" {
-            val agent = makeParserAgent()
-            val response = "<toolName>Answer</toolName>"
-
-            val (intention, toolName) = agent.parseIntentionAndTool(response, validTools)
-
-            intention shouldBe response.trim()
-            toolName shouldBe "Answer"
-        }
-
-        "parseIntentionAndTool — completely empty response throws IntentionParsingException" {
-            val agent = makeParserAgent()
-
-            shouldThrow<AgentAdvanced.IntentionParsingException> {
-                agent.parseIntentionAndTool("", validTools)
-            }
-        }
-
-        "parseIntentionAndTool — tool name matching is case-insensitive" {
-            val agent = makeParserAgent()
-            val response =
-                """
-                <intention>Reading the file.</intention>
-                <toolName>files__readfile</toolName>
-                """.trimIndent()
-
-            val (_, toolName) = agent.parseIntentionAndTool(response, validTools)
-
-            toolName shouldBe "FILES__ReadFile"
-        }
-
+        fun makeInitialEvents(
+            namespaceId: UUID,
+            caseId: UUID,
+        ) = listOf(
+            MessageEvent(
+                namespaceId = namespaceId,
+                caseId = caseId,
+                actor = Actor("user1", "User One", ActorRole.USER),
+                content = listOf(MessageContent.Text("Hello, can you help me?")),
+            ),
+        )
         // -------------------------------------------------------------------------
         // Redirect contract
         // -------------------------------------------------------------------------
@@ -169,22 +70,35 @@ class AgentAdvancedSpec :
 
             val mockChatClient = mockk<ChatClient>(relaxed = true)
             val mockStreamSpec = mockk<ChatClient.StreamResponseSpec>(relaxed = true)
-            // First call() → intention selects the redirect tool
-            // Second call() → parameter generation returns valid JSON for the tool
+            // call() → parameter generation returns valid JSON for the redirect tool
             every {
                 mockChatClient.prompt(any<Prompt>()).call().content()
-            } returnsMany listOf(
-                "<intention>Redirect to TargetAgent.</intention><toolName>REDIRECT__redirect</toolName>",
-                """{"agentName":"TargetAgent"}""",
-            )
+            } returns """{"agentName":"TargetAgent"}"""
             every { mockChatClient.prompt(any<Prompt>()).stream() } returns mockStreamSpec
             every { mockStreamSpec.content() } returns Flux.empty()
 
+            val mockGenerator = mockk<AgentIntentionGenerator>()
+            every {
+                mockGenerator.generate(any(), any(), any(), any(), any())
+            } returns IntentionGeneratedEvent(
+                namespaceId = namespaceId,
+                caseId = caseId,
+                agentId = agentId,
+                intention = "Redirect to TargetAgent.",
+                toolName = "REDIRECT__redirect",
+            )
+
+            val context = AgentAdvancedContext(
+                chatClient = mockChatClient,
+                tools = listOf(redirectTool),
+                instructions = null,
+                agentId = agentId,
+            )
             val agent = AgentAdvanced(
                 metadata = EntityMetadata(id = agentId),
                 name = "TestAgent",
-                chatClient = mockChatClient,
-                tools = listOf(redirectTool),
+                context = context,
+                intentionGenerator = mockGenerator,
                 maxIterations = 5,
             )
 
@@ -222,20 +136,35 @@ class AgentAdvancedSpec :
 
             val mockChatClient = mockk<ChatClient>(relaxed = true)
             val mockStreamSpec = mockk<ChatClient.StreamResponseSpec>(relaxed = true)
+            // call() → parameter generation returns valid JSON for the redirect tool
             every {
                 mockChatClient.prompt(any<Prompt>()).call().content()
-            } returnsMany listOf(
-                "<intention>Redirect to TargetAgent.</intention><toolName>REDIRECT__redirect</toolName>",
-                """{"agentName":"TargetAgent"}""",
-            )
+            } returns """{"agentName":"TargetAgent"}"""
             every { mockChatClient.prompt(any<Prompt>()).stream() } returns mockStreamSpec
             every { mockStreamSpec.content() } returns Flux.empty()
 
+            val mockGenerator = mockk<AgentIntentionGenerator>()
+            every {
+                mockGenerator.generate(any(), any(), any(), any(), any())
+            } returns IntentionGeneratedEvent(
+                namespaceId = namespaceId,
+                caseId = caseId,
+                agentId = agentId,
+                intention = "Redirect to TargetAgent.",
+                toolName = "REDIRECT__redirect",
+            )
+
+            val context = AgentAdvancedContext(
+                chatClient = mockChatClient,
+                tools = listOf(redirectTool),
+                instructions = null,
+                agentId = agentId,
+            )
             val agent = AgentAdvanced(
                 metadata = EntityMetadata(id = agentId),
                 name = "TestAgent",
-                chatClient = mockChatClient,
-                tools = listOf(redirectTool),
+                context = context,
+                intentionGenerator = mockGenerator,
                 maxIterations = 5,
             )
 
@@ -253,43 +182,46 @@ class AgentAdvancedSpec :
             events.filterIsInstance<WarnEvent>() shouldHaveSize 0
         }
 
+
         // -------------------------------------------------------------------------
         // Orchestration integration tests
         // -------------------------------------------------------------------------
-
-        fun makeInitialEvents(
-            namespaceId: UUID,
-            caseId: UUID,
-        ) = listOf(
-            MessageEvent(
-                namespaceId = namespaceId,
-                caseId = caseId,
-                actor = Actor("user1", "User One", ActorRole.USER),
-                content = listOf(MessageContent.Text("Hello, can you help me?")),
-            ),
-        )
 
         "should complete full orchestration loop when LLM returns Answer tool in XML" {
             val namespaceId = UUID.randomUUID()
             val caseId = UUID.randomUUID()
             val agentId = UUID.randomUUID()
 
+            val mockGenerator = mockk<AgentIntentionGenerator>()
+            every {
+                mockGenerator.generate(any(), any(), any(), any(), any())
+            } returns
+                IntentionGeneratedEvent(
+                    namespaceId = namespaceId,
+                    caseId = caseId,
+                    agentId = agentId,
+                    intention = "No further tool calls needed.",
+                    toolName = "Answer",
+                )
+
             val mockChatClient = mockk<ChatClient>(relaxed = true)
             val mockStreamSpec = mockk<ChatClient.StreamResponseSpec>(relaxed = true)
-
-            // call() is used for intention generation; stream() is used for the final answer
-            every {
-                mockChatClient.prompt(any<Prompt>()).call().content()
-            } returns "<intention>No further tool calls needed.</intention><toolName>Answer</toolName>"
             every { mockChatClient.prompt(any<Prompt>()).stream() } returns mockStreamSpec
             every { mockStreamSpec.content() } returns Flux.just("Here is ", "my answer.")
 
+            val context =
+                AgentAdvancedContext(
+                    chatClient = mockChatClient,
+                    tools = emptyList(),
+                    instructions = null,
+                    agentId = agentId,
+                )
             val agent =
                 AgentAdvanced(
                     metadata = EntityMetadata(id = agentId),
                     name = "TestAgent",
-                    chatClient = mockChatClient,
-                    tools = emptyList(),
+                    context = context,
+                    intentionGenerator = mockGenerator,
                     maxIterations = 5,
                 )
 
@@ -329,97 +261,91 @@ class AgentAdvancedSpec :
             finishedEvent.agentName shouldBe "TestAgent"
         }
 
-        "resolveIntentionAndTool retries on malformed LLM response and succeeds on second attempt" {
+        "emits WarnEvent when repetition loop is detected and agent eventually selects Answer" {
             val namespaceId = UUID.randomUUID()
             val caseId = UUID.randomUUID()
 
             val mockChatClient = mockk<ChatClient>(relaxed = true)
             val mockStreamSpec = mockk<ChatClient.StreamResponseSpec>(relaxed = true)
+            every { mockChatClient.prompt(any<Prompt>()).stream() } returns mockStreamSpec
+            every { mockStreamSpec.content() } returns Flux.just("Loop stopped.")
 
+            val mockTool = mockk<io.whozoss.agentos.sdk.tool.StandardTool<String>>(relaxed = true)
+            every { mockTool.name } returns "FILES__ReadFile"
+            every { mockTool.description } returns "Read a file"
+            every { mockTool.inputSchema } returns "{}"
+            every { mockTool.paramType } returns String::class.java
+            every { mockTool.executeWithJson(any(), any()) } returns "file content"
+
+            // The generator returns FILES__ReadFile 3 times, then Answer on the 4th call
+            val mockGenerator = mockk<AgentIntentionGenerator>()
             every {
-                mockChatClient.prompt(any<Prompt>()).call().content()
+                mockGenerator.generate(any(), any(), any(), any(), isNull())
             } returnsMany
                 listOf(
-                    "This is a malformed response with no XML tags at all",
-                    "<intention>All good on retry.</intention><toolName>Answer</toolName>",
+                    IntentionGeneratedEvent(
+                        namespaceId = namespaceId,
+                        caseId = caseId,
+                        agentId = UUID.randomUUID(),
+                        intention = "Reading the file.",
+                        toolName = "FILES__ReadFile",
+                    ),
+                    IntentionGeneratedEvent(
+                        namespaceId = namespaceId,
+                        caseId = caseId,
+                        agentId = UUID.randomUUID(),
+                        intention = "Reading the file again.",
+                        toolName = "FILES__ReadFile",
+                    ),
+                    IntentionGeneratedEvent(
+                        namespaceId = namespaceId,
+                        caseId = caseId,
+                        agentId = UUID.randomUUID(),
+                        intention = "Reading the file yet again.",
+                        toolName = "FILES__ReadFile",
+                    ),
                 )
-            every { mockChatClient.prompt(any<Prompt>()).stream() } returns mockStreamSpec
-            every { mockStreamSpec.content() } returns Flux.just("Done.")
-
-            val agent =
-                AgentAdvanced(
-                    name = "RetryAgent",
-                    chatClient = mockChatClient,
-                    tools = emptyList(),
+            every {
+                mockGenerator.generate(any(), any(), any(), any(), not(isNull()))
+            } returns
+                IntentionGeneratedEvent(
+                    namespaceId = namespaceId,
+                    caseId = caseId,
+                    agentId = UUID.randomUUID(),
+                    intention = "Stopping due to loop.",
+                    toolName = "Answer",
                 )
 
-            val events = agent.run(makeInitialEvents(namespaceId, caseId)).toList()
-
-            val intentionEvent = events.filterIsInstance<IntentionGeneratedEvent>().first()
-            intentionEvent.toolName shouldBe "Answer"
-            intentionEvent.intention shouldContain "All good on retry"
-        }
-
-        "resolveIntentionAndTool retries on LLM service failure and succeeds on second attempt" {
-            val namespaceId = UUID.randomUUID()
-            val caseId = UUID.randomUUID()
-
-            val mockChatClient = mockk<ChatClient>(relaxed = true)
-            val mockStreamSpec = mockk<ChatClient.StreamResponseSpec>(relaxed = true)
-
-            var callCount = 0
+            // Also mock call() for generateParameters
             every {
                 mockChatClient.prompt(any<Prompt>()).call().content()
-            } answers {
-                callCount++
-                if (callCount == 1) throw RuntimeException("Service unavailable")
-                "<intention>Recovered after failure.</intention><toolName>Answer</toolName>"
-            }
-            every { mockChatClient.prompt(any<Prompt>()).stream() } returns mockStreamSpec
-            every { mockStreamSpec.content() } returns Flux.just("Done.")
+            } returns "{}"
 
+            val agentId = UUID.randomUUID()
+            val context =
+                AgentAdvancedContext(
+                    chatClient = mockChatClient,
+                    tools = listOf(mockTool),
+                    instructions = null,
+                    agentId = agentId,
+                )
             val agent =
                 AgentAdvanced(
-                    name = "RetryAgent",
-                    chatClient = mockChatClient,
-                    tools = emptyList(),
+                    metadata = EntityMetadata(id = agentId),
+                    name = "LoopAgent",
+                    context = context,
+                    intentionGenerator = mockGenerator,
+                    maxIterations = 10,
                 )
 
             val events = agent.run(makeInitialEvents(namespaceId, caseId)).toList()
 
-            val intentionEvent = events.filterIsInstance<IntentionGeneratedEvent>().first()
-            intentionEvent.toolName shouldBe "Answer"
-            intentionEvent.intention shouldContain "Recovered after failure"
-        }
+            val warnEvents = events.filterIsInstance<WarnEvent>()
+            warnEvents.any { it.message.contains("consecutively") || it.message.contains("FILES__ReadFile") } shouldBe true
 
-        "resolveIntentionAndTool falls back to Answer after all retry attempts exhausted" {
-            val namespaceId = UUID.randomUUID()
-            val caseId = UUID.randomUUID()
+            val intentionEvents = events.filterIsInstance<IntentionGeneratedEvent>()
+            intentionEvents.last().toolName shouldBe "Answer"
 
-            val mockChatClient = mockk<ChatClient>(relaxed = true)
-            val mockStreamSpec = mockk<ChatClient.StreamResponseSpec>(relaxed = true)
-
-            every {
-                mockChatClient.prompt(any<Prompt>()).call().content()
-            } returns "This is always malformed with no XML tags"
-            every { mockChatClient.prompt(any<Prompt>()).stream() } returns mockStreamSpec
-            every { mockStreamSpec.content() } returns Flux.just("Fallback answer.")
-
-            val agent =
-                AgentAdvanced(
-                    name = "FallbackAgent",
-                    chatClient = mockChatClient,
-                    tools = emptyList(),
-                )
-
-            val events = agent.run(makeInitialEvents(namespaceId, caseId)).toList()
-
-            val intentionEvent = events.filterIsInstance<IntentionGeneratedEvent>().first()
-            intentionEvent.toolName shouldBe "Answer"
-            // intention should be the raw last response (trimmed)
-            intentionEvent.intention shouldBe "This is always malformed with no XML tags"
-
-            // Agent should still finish normally
             events.filterIsInstance<AgentFinishedEvent>() shouldHaveSize 1
         }
 
@@ -519,90 +445,5 @@ class AgentAdvancedSpec :
                 )
 
             agent.detectRepetitionLoop(events) shouldBe null
-        }
-
-        "emits WarnEvent when repetition loop is detected and agent eventually selects Answer" {
-            val namespaceId = UUID.randomUUID()
-            val caseId = UUID.randomUUID()
-
-            val mockChatClient = mockk<ChatClient>(relaxed = true)
-            val mockStreamSpec = mockk<ChatClient.StreamResponseSpec>(relaxed = true)
-
-            // Each tool iteration consumes 2 call().content() invocations:
-            //   1st: resolveIntentionAndTool → returns toolName
-            //   2nd: generateParameters → returns JSON args (any string works)
-            // After 3 tool iterations (6 calls), the 7th call is the 4th intention → Answer.
-            every {
-                mockChatClient.prompt(any<Prompt>()).call().content()
-            } returnsMany
-                listOf(
-                    "<intention>Reading the file.</intention><toolName>FILES__ReadFile</toolName>", // iter 1 intention
-                    "{}", // iter 1 params
-                    "<intention>Reading the file again.</intention><toolName>FILES__ReadFile</toolName>", // iter 2 intention
-                    "{}", // iter 2 params
-                    "<intention>Reading the file yet again.</intention><toolName>FILES__ReadFile</toolName>", // iter 3 intention
-                    "{}", // iter 3 params
-                    "<intention>Stopping due to loop.</intention><toolName>Answer</toolName>", // iter 4 intention (after WarnEvent)
-                )
-            every { mockChatClient.prompt(any<Prompt>()).stream() } returns mockStreamSpec
-            every { mockStreamSpec.content() } returns Flux.just("Loop stopped.")
-
-            val mockTool = mockk<io.whozoss.agentos.sdk.tool.StandardTool<String>>(relaxed = true)
-            every { mockTool.name } returns "FILES__ReadFile"
-            every { mockTool.description } returns "Read a file"
-            every { mockTool.inputSchema } returns "{}"
-            every { mockTool.paramType } returns String::class.java
-            every { mockTool.executeWithJson(any(), any()) } returns "file content"
-
-            val agent =
-                AgentAdvanced(
-                    name = "LoopAgent",
-                    chatClient = mockChatClient,
-                    tools = listOf(mockTool),
-                    maxIterations = 10,
-                )
-
-            val events = agent.run(makeInitialEvents(namespaceId, caseId)).toList()
-
-            val warnEvents = events.filterIsInstance<WarnEvent>()
-            warnEvents shouldHaveSize 1
-            warnEvents.first().message shouldContain "FILES__ReadFile"
-            warnEvents.first().message shouldContain "consecutively"
-
-            val intentionEvents = events.filterIsInstance<IntentionGeneratedEvent>()
-            intentionEvents.last().toolName shouldBe "Answer"
-
-            events.filterIsInstance<AgentFinishedEvent>() shouldHaveSize 1
-        }
-
-        "resolveIntentionAndTool falls back to Answer after repeated LLM failures" {
-            val namespaceId = UUID.randomUUID()
-            val caseId = UUID.randomUUID()
-
-            val mockChatClient = mockk<ChatClient>(relaxed = true)
-            val mockStreamSpec = mockk<ChatClient.StreamResponseSpec>(relaxed = true)
-
-            every {
-                mockChatClient.prompt(any<Prompt>()).call().content()
-            } throws RuntimeException("Persistent service failure")
-            every { mockChatClient.prompt(any<Prompt>()).stream() } returns mockStreamSpec
-            every { mockStreamSpec.content() } returns Flux.just("Fallback answer.")
-
-            val agent =
-                AgentAdvanced(
-                    name = "FallbackAgent",
-                    chatClient = mockChatClient,
-                    tools = emptyList(),
-                )
-
-            val events = agent.run(makeInitialEvents(namespaceId, caseId)).toList()
-
-            val intentionEvent = events.filterIsInstance<IntentionGeneratedEvent>().first()
-            intentionEvent.toolName shouldBe "Answer"
-            // No lastResponse captured since all calls threw — falls back to default message
-            intentionEvent.intention shouldBe "Unable to generate intention"
-
-            // Agent should still finish normally
-            events.filterIsInstance<AgentFinishedEvent>() shouldHaveSize 1
         }
     })
