+++ agentos/agentos-service/src/test/kotlin/io/whozoss/agentos/agent/AgentServiceImplUnitSpec.kt
@@ -36,6 +36,7 @@ class AgentServiceImplUnitSpec : StringSpec() {
     private val integrationConfigService: IntegrationConfigService = mockk(relaxed = true)
     private val userService: UserService = mockk(relaxed = true)
     private val agentConfigService: AgentConfigService = mockk()
+    private val intentionGenerator: AgentIntentionGenerator = mockk(relaxed = true)
     private val agentService =
         AgentServiceImpl(
             chatClientProvider,
@@ -46,6 +47,7 @@ class AgentServiceImplUnitSpec : StringSpec() {
             integrationConfigService,
             userService,
             agentConfigService,
+            intentionGenerator,
         )
 
     private val namespaceId: UUID = UUID.randomUUID()
@@ -290,6 +292,7 @@ class AgentServiceImplUnitSpec : StringSpec() {
                     localIntegrationService,
                     userService,
                     agentConfigService,
+                    intentionGenerator,
                 )
             val configs =
                 listOf(
